# dialect/sql/sqlbuild: replace bytes.buffer with strings.builder